Incorporating feedback about Challenger #1
# Challenge 1 feedback

Hi Patrick,

Your grade for Challenge 1 is: 90/100

Some feedback from Derek, me and Kimberly: Well done. Great presentation and excellent visually. Some of the content on CASs (especially that from popular Wiki pages is a bit inaccurate or at least misleading like the whole being greater than the sum, aggregate). You could cite others and use a different quote that is less misleading. To the general public it will go unnoticed but to CAS folks it could rub them the wrong way. We like the break down and the ability to interact and test it out. It is a great-use of modeling/code to make an interactive site.

Let me know if you have any questions. Good job!

Have a great weekend!
Laura

---

- Added less misleading citation by Holland (2006).
- Fixes for text that was affected by that citation